requirements.txt
# Veri Çekme
yfinance         # Yahoo Finance API'den veri çekmek için

# Veri İşleme
pandas           # Veri manipülasyonu ve analiz için
numpy            # Sayısal işlemler için
pyspark          # Büyük veri işlemleri için PySpark
sqlalchemy       # Veritabanı bağlantısı ve işlemleri için

# Jupyter Notebook işlemleri için
notebook

# Makine Öğrenmesi ve Modelleme
scikit-learn     # Makine öğrenmesi modelleri (örneğin, Random Forest) için
xgboost          # XGBoost modeli için
prophet          # Prophet ile zaman serisi tahmini için
statsmodels      # Zaman serisi modelleme (ARIMA gibi) için

# Veritabanı İşlemleri
psycopg2-binary  # PostgreSQL veritabanı bağlantısı için

# Görselleştirme
matplotlib       # Veri görselleştirme için
seaborn          # Grafik ve görselleştirme desteği için

# Ek Araçlar
tqdm             # İlerleme çubuğu (ETL veya uzun süren işlemler için faydalı)
pyyaml           # YAML dosyalarıyla çalışma (örneğin, config dosyaları için)
